v0.10.1 - Upravená databáze pro záznam kontroly, celý layout pro přidání dodělán. Upravena grafika a rozložení Informačního fragmentu pro stanoviste

+++ ulyDetailModule/src/main/java/com/hruby/ulydetailmodule/ui/namereneHodnotyUl/NamereneHodnotyUlRecycleViewAdapter.kt
@@ -37,7 +37,7 @@ class NamereneHodnotyUlRecycleViewAdapter(
 
     override fun onBindViewHolder(holder: NamereneHodnotyViewHolder, position: Int) {
         val hodnoty = hodnotyList[position]
-        holder.datetime.text = formatTimestamp(hodnoty.datum)
+        holder.datetime.text = formatTimestamp(hodnoty.datum.toLong())
         holder.tempUl.text = String.format(hodnoty.teplotaUl.toString())
         holder.tempModul.text = String.format(hodnoty.teplotaModul.toString())
         holder.humModul.text = String.format(hodnoty.vlhkostModul.toString())
